shard_id_t: change NO_SHARD to sort before 0 (min instead of max)
The min ghobject has shard NO_SHARD, and is the default constructed value.
That initial value is assumed in uncounted ways across the code base when
users do

  ghobject_t foo;
  foo.this = that;

such that changing it is dangerous.  It is safer to change the shard_id_t
sort order such that NO_SHARD is signed instead of unsigned.  The value
doesn't actually change (still 0xff), but the sorting does.  Note that
only a single comparison triggers a signed/unsigned warning from this
change, and it assumes that the shard is not NO_SHARD (ec pool) and we
case it to preserve the old behavior anyway.

In PGBackend we change the minimum value for the objects_list_partial()
method to start with NO_SHARD.

Signed-off-by: Sage Weil <sage@redhat.com>