mon: return error on forwarded "tell version" command
For not a leader, if a received monitor command is not supported
locally, but is supported by the leader, it is forwarded to the
leader. For "ceph tell mon.x version" it gives a confusing behavior:
if the mon.x is not a leader and does not support "version" command
yet, but the leader does, the user will receive the version of the
leader, so the user can't be actually sure about a non leader version.

Fix this by checking if the received "version" command is forwarded
and return error in this case.

Signed-off-by: Mykola Golub <mgolub@mirantis.com>