mon: 'tell' availability for monitor commands
For not a monitor leader, if a received command is not supported
locally, but is supported by the leader, it is forwarded to the
leader.

For 'ceph tell mon.x' commands that has undesirable behaviour: if the
mon.x is not a leader and does not support the command yet, but the
leader does, the user will receive a reponse from the leader.

Fix this by adding 'tell' availability.

Signed-off-by: Mykola Golub <mgolub@mirantis.com>