ceph.spec.in: Add -DFMT_DEPRECATED_OSTREAM to CXXFLAGS
This is a temporary downstream fix so that Ceph will continue to build on
openSUSE Factory, which is updating to fmt version 9.  Ceph currently fails
to build when using fmt 9 (see https://tracker.ceph.com/issues/56610), so
this fix will unstick us on Factory until we can get a proper fix upstream.
Note that per fmt changelog, the FMT_DEPRECATED_OSTREAM define will be
removed in the next major release, but we'll be good with this immediately
for the time being.  See:
https://github.com/fmtlib/fmt/blob/master/ChangeLog.rst#900---2022-07-04

Fixes: https://bugzilla.opensuse.org/show_bug.cgi?id=1202292
Signed-off-by: Tim Serong <tserong@suse.com>